+++ projects/vr-lib/src/public-api.ts
@@ -3,9 +3,11 @@
  */
 
 export { VrKitHeaderComponent } from './lib/components/header/header.component';
+export type { NavigationItem, UserInfoModel } from './lib/components/header/header.component';
 export { VrKitPageHeaderComponent } from './lib/components/page-header/page-header.component';
 export { VrKitBreadcrumbComponent } from './lib/components/breadcrumb/breadcrumb.component';
 export { VrKitArticleComponent } from './lib/components/article/article.component';
 export { VrKitStepperComponent } from './lib/components/stepper/stepper.component';
 export { VrKitIconComponent } from './lib/components/icon/icon.component';
 export { VrKitNoDataComponent } from './lib/components/no-data/no-data.component';
+export { ThemeService, DAY_THEME, NIGHT_THEME, VR_THEME, THEME_KEY } from './lib/services/theme.service';